Offer to autofix constant and method visibility by adding public keyword.
Public was chosen as that's what PHP interprets the lack of a visibility modifier as.
This will remove the code smell, but may introduce another in that the member is now
public, but may only be used internally. Human review is required to convert public
down to a more restrictive value, doing so may also introduce php error situations
if a child class is still defaulted into public, but a parent class has now been locked
down to somethign more restrictive.